QoL updates to CONTRIBUTING doc
The URL in the env variables for getting an oauth access token was
incorrect. I believe it's set properly on prod, as oauth is working
properly in the production Slack app. But the `/oauth` endpoint doesn't
exist in CWA, and associating with a data.world account didn't work
unless I changed it to `/oauth/access_token`.

Also added a semicolon in the Postgres instructions to automatically end
the line without user input.