Updates to CONTRIBUTING doc (#42)
* QoL updates to CONTRIBUTING doc

The URL in the env variables for getting an oauth access token was
incorrect. I believe it's set properly on prod, as oauth is working
properly in the production Slack app. But the `/oauth` endpoint doesn't
exist in CWA, and associating with a data.world account didn't work
unless I changed it to `/oauth/access_token`.

Also added a semicolon in the Postgres instructions to automatically end
the line without user input.

* link to create classic Slack app

Our Slack app was created as an old Slack app and currently has some
compatibility issues when used with a new Slack app. This should be
updated once our Slack app is migrated to the new version.