Add command for generating per-channel webhook URLs (#43)
* add webhookId column to Channels table in db

I updated the config to include the dialect at the top level because I
was getting an error `Dialect needs to be explicitly supplied as of
v4.0.0` while running migrations. Adding this line to the config fixed
it.

* add webhook Slack command

Running `/data.world webhook` gives the user a webhook id for the
current channel. This webhook id is generated if there is no associated
id in the database yet.

I created a new directory `server/commands/` for files related to
commands. The `server/commands/command.js` was getting getting quite
large, determining which commands should be run while also storing all
of the logic that the command should perform. Right now I've put the
`webhook` command logic in this directory - ideally the `subscribe`
command logic would be moved there in the future.

* edit help message to include webhook command

* add tests for webhook command

* address PR feedback