Handle webhook request events (#44)
* move commands directory under controllers

I felt commands didn't make sense as a top-level directory. Since all of
the commands would be used in the controllers, I moved it as a
sub-directory under controllers/

* get Slack blocks message from webhook body

* use new Slack SDK for sending messages with blocks

The old SDK used before did not support sending messages with the new
block format. Adding the new Slack SDK as a dependency and using both in
tandem for now, until we get a chance to clean the old one up.

* create helpers for getting bot access tokens

* tests for new webhook endpoint

* use correct origin in formatted urls

For the URLs that are built based on agentid/datasetid, we should
probably use the same origin as the URLs that are sent in the webhook
request.

* add tests for request event Slack block creation

* explicitly require URL dependency

The CircleCI environment is still using Node 8. So we have to explicitly
require this module in order to use it.

https://www.nodebeginner.org/blog/post/nodejs-tutorial-whatwg-url-parser/

* address PR feedback

* import collections in event handlers